refactor: remove logic from form
