+++ app/src/components/sections/SingleTeamView/Team/MemberCard.vue
@@ -47,10 +47,13 @@
         from the team?
       </DeleteConfirmForm>
     </ModalComponent>
-    <ModalComponent v-model="showAddEditMemberRoles">
+    <ModalComponent v-if="roleCategories" v-model="showAddEditMemberRoles">
       <AddMemberRolesForm
-        :member-address="member.address"
+        :role-categories="roleCategories"
+        :is-adding-role="isAddingRole"
         v-model="member.roles"
+        @add-roles="addRoles"
+        @close-modal="showAddEditMemberRoles = !showAddEditMemberRoles"
       />
     </ModalComponent>
   </div>
@@ -60,14 +63,15 @@ import { useUserDataStore } from '@/stores/user'
 import DeleteConfirmForm from '@/components/forms/DeleteConfirmForm.vue'
 import ModalComponent from '@/components/ModalComponent.vue'
 import { useRoute } from 'vue-router'
-import type { MemberInput } from '@/types'
+import type { MemberInput, RoleCategory, Role } from '@/types'
 import { useClipboard } from '@vueuse/core'
 import { NETWORK } from '@/constant'
-import { ref, watch } from 'vue'
+import { ref, watch, onMounted } from 'vue'
 import { useErrorHandler } from '@/composables/errorHandler'
 import { useToastStore } from '@/stores/useToastStore'
 import { useCustomFetch } from '@/composables/useCustomFetch'
 import AddMemberRolesForm from '@/components/sections/SingleTeamView/Team/forms/AddMemberRolesForm.vue'
+import { log, parseError } from "@/utils";
 
 const props = defineProps<{
   member: Partial<MemberInput>
@@ -83,6 +87,8 @@ const route = useRoute()
 const memberToBeDeleted = ref({ name: '', address: '', id: '' })
 const showDeleteMemberConfirmModal = ref(false)
 const showAddEditMemberRoles = ref(false)
+const roleCategories = ref<null | RoleCategory[]>(null)
+const isAddingRole = ref(false)
 
 // useFetch instance for deleting member
 const {
@@ -101,6 +107,16 @@ const {
   }
 })
   .delete()
+  .json()
+
+// useFetch fetch role categories
+const {
+  execute: executeFetchRoleCategories,
+  data: _roleCategories
+} = useCustomFetch('role-category', {
+  immediate: false
+})
+  .get()
   .json()
 // Watchers for deleting member
 watch([() => memberIsDeleting.value, () => deleteMemberError.value], async () => {
@@ -123,4 +139,109 @@ const { copy, copied, isSupported } = useClipboard()
 const openExplorer = (address: string) => {
   window.open(`${NETWORK.blockExplorerUrl}/address/${address}`, '_blank')
 }
+
+const createContract = () => {
+  let contract
+  if (member.value.roles)
+    for (const memberRole of member.value.roles) {
+      const roleCategory = _roleCategories
+        .value
+        .roleCategories
+        .find((category: RoleCategory) => 
+          category.id === (memberRole as any).role.roleCategoryId)
+
+      if (roleCategory && roleCategory.roles) {
+        const role = roleCategory
+          .roles
+          .find(
+            (_role: Role) => 
+              _role.id === (memberRole as any).roleId
+          )
+
+        const entitlements = []
+
+        if (role && role.entitlements) {
+          for (const entitlement of role.entitlements) {
+            if (
+              entitlement.entitlementType &&
+              entitlement.entitlementType.name === 'access' &&
+              entitlement.value.split(':')[0] === 'expense-account'
+            ) {
+              entitlements.push(entitlement.value)
+            }
+          }
+
+          if (entitlements.length > 0) {
+            contract = {
+              role: {
+                name: role.name,
+                entitlement: {
+                  name: "access",
+                  resource: entitlements[0].split(':')[0],
+                  accessLevel: entitlements[0].split(':')[1]
+                }
+              },
+              assignedTo: member.value.address,
+              assignedBy: useUserDataStore().address
+            }
+          }
+        }
+      }
+    }
+  return contract
+}
+
+const signContract = async () => {
+  if (!createContract()) return
+  const params = [
+    useUserDataStore().address,
+    {
+      types: {
+        EIP712Domain: [
+          { name: "name", type: "string" },
+          { name: "version", type: "string" }
+        ],
+        Entitlement: [
+          { name: "name", type: "string" },
+          { name: "resource", type: "string" },
+          { name: "accessLevel", type: "string" }
+        ],
+        Role: [
+          { name: "name", type: "string" },
+          { name: "entitlement", type: "Entitlement" }
+        ],
+        Contract: [
+          { name: "assignedTo", type: "address" },
+          { name: "assignedBy", type: "address" },
+          { name: "role", type: "Role" }
+        ]
+      },
+      primaryType: "Contract",
+      domain: {
+        "name": "CNC Contract",
+        "version": "1"
+      },
+      message: createContract()
+    }
+  ]
+  try {
+    return await (window as any).ethereum.request({method: "eth_signTypedData_v4", params: params})
+  } catch (error) {
+    log.error(parseError(error))
+  }
+}
+
+const addRoles = async () => {
+  isAddingRole.value = true
+  const signature = await signContract()
+  console.log(`member.roles: `, member.value.roles)
+  console.log(`signature: `, signature)
+  console.log(`contract: `, JSON.stringify(createContract()))
+  isAddingRole.value = false
+}
+
+onMounted(async () => {
+ await executeFetchRoleCategories()
+ roleCategories.value = _roleCategories.value.roleCategories
+})
 </script>
